+++ js/utils/get-random-positive-float.js
@@ -1,6 +1,6 @@
 // Функция взята из интернета и доработана
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
-
+/*
 function getRandomPositiveFloat (a, b, digits = 1) {
   // Чтобы не заставлять пользователя нашей функции помнить порядок аргументов,
   // реализуем поддержку передачи минимального и максимального значения в любом порядке,
@@ -19,4 +19,4 @@ function getRandomPositiveFloat (a, b, digits = 1) {
   // И в конце с помощью метода toFixed любого числа в JavaScript
   // указать требуемое количество знаков после точки
   return result.toFixed(digits);
-}
+}*/
